Create Stock Price Prediction with Prophet Model.py
Project Description
This project is part of my thesis research. This project aims to build a Prophet model that matches the characteristics of stock price data. The data used is the closing price data of BIRD.JK shares. The Prophet model (basic) in the Python package is a deterministic model, while stock prices are data with a stochastic process. Therefore, in this study, the Prophet model (basic) is modified by adding a normally distributed random number component with parameters obtained from the error value of the Prophet model (basic). 

Methodology:
1. Descriptive analysis for data exploration,
2. Data division into training and testing,
3. Building the basic Prophet model with training data,
4. Modifying the Prophet model,
5. Determining the prediction results with the modified Prophet model,
6. Determining the model accuracy level.

Final Results
The prediction results of the modified Prophet model vary because they are stochastic. This differs from the base Prophet model, which produces the same results in every computation process. The prediction results of the modified Prophet model can be better than those of the base Prophet model.