Create Create Stock Price Prediction with Markov Switching - Autoregressive Model.py
This project is part of my thesis research. This project aims to build a Markov Switching - Autoregressive (MS-AR) model that matches the characteristics of stock price data. The data used is the closing price data of BIRD.JK shares.

Methodology:
1. Descriptive analysis for data exploration,
2. Data division into training and testing,
3. Building MS-AR model,
5. Determining the prediction results with the MS-AR model,
6. Determining the model accuracy level.

Final Result:
The MS-AR model is capable of accurately predicting the price of BIRD.JK shares.